Move Json test cases into non-test file so Dolt can depend on them.
